requirements.txt
django
djangorestframework
psycopg2-binary
django-cors-headers
django-extensions
django-redis
celery[redis]

